+++ src/main/java/framework/buttons/ButtonRevealQueue.java
@@ -3,13 +3,11 @@ package framework.buttons;
 import abr.queue_abr.queue.QueueGetInputBoundary;
 import abr.queue_abr.queue.QueueGetOutputBoundary;
 import abr.queue_abr.queue.QueueGetUseCase;
-import interface_adaptors.SearchResultsViewModel;
+import interface_adaptors.search_engine_ia.SearchResultsViewModel;
 import interface_adaptors.playlist_ia.RecordViewModel;
 import interface_adaptors.queue_ia.QueueGetController;
 import interface_adaptors.queue_ia.QueueGetPresenter;
 import interface_adaptors.queue_ia.QueueViewModel;
-import interface_adaptors.song_player_ia.SongPlayerController;
-import interface_adaptors.song_player_ia.SongPlayerViewModel;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
