+++ src/main/java/interface_adaptors/display_ia/DisplaySearchUseCase.java
@@ -1,12 +1,7 @@
 package interface_adaptors.display_ia;
 
-import entities.Song;
-import entities.user_entities.User;
 import interface_adaptors.AbstractDisplayUseCase;
-import interface_adaptors.SearchBarViewModel;
-import interface_adaptors.SearchResultsViewModel;
-
-import java.util.ArrayList;
+import interface_adaptors.search_engine_ia.SearchBarViewModel;
 
 public class DisplaySearchUseCase extends AbstractDisplayUseCase {
