+++ src/main/java/interface_adaptors/playlist_ia/PlaylistCollectionViewModel.java
@@ -1,14 +1,13 @@
-package interface_adaptors;
+package interface_adaptors.playlist_ia;
 
-import entities.playlist_entities.Playlist;
 import framework.buttons.ButtonCreatePlaylist;
 import framework.items.PlaylistCollectionItem;
+import interface_adaptors.AbstractViewModel;
 import interface_adaptors.user_login_ia.UserStatusObserver;
 import interface_adaptors.user_login_ia.UserStatusViewModel;
 
 import javax.swing.*;
 import java.awt.*;
-import java.util.ArrayList;
 import java.util.List;
 
 public class PlaylistCollectionViewModel extends AbstractViewModel<List<String>> implements UserStatusObserver {
